requirements.txt
# google-adk==1.7.0
# google-cloud-aiplatform[adk,agent_engines]==1.60.0
# pydantic==2.9.0
# python-dotenv==1.0.1
# cloudpickle==3.1.0 
#

google-adk>=1.7.0,<2.0.0
google-cloud-aiplatform[adk,agent_engines]>=1.49.0,<2.0.0
pydantic>=2.11.3,<3.0.0
python-dotenv>=1.1.0,<2.0.0
requests>=2.31.0,<3.0.0
cloudpickle>=3.1.0,<4.0.0




# requests>=2.31.0,<3.0.0
# google-adk>=1.7.0,<2.0.0
# pydantic>=2.11.3,<3.0.0
# python-dotenv>=1.1.0,<2.0.0
# google-cloud-aiplatform[adk,agent_engines]>=1.42.1,<2.0.0
# absl-py>=2.1.0,<3.0.0
# cloudpickle>=3.0.0,<4.0.0